pyproject.toml
[tool.poetry]
name = "bank-transaction-categorizer"
version = "0.1.0"
description = "Categorize bank transactions and identify vendors"
authors = ["Developer"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
streamlit = ">=1.49.1,<2.0.0"
pandas = ">=2.3.2,<3.0.0"
sqlalchemy = ">=2.0.43,<3.0.0"
openai = ">=1.107.2,<2.0.0"
anthropic = ">=0.40.0,<1.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
pydantic = ">=2.11.9,<3.0.0"
celery = ">=5.3.0,<6.0.0"
redis = ">=5.0.0,<6.0.0"
plotly = "^5.17.0"
seaborn = "^0.13.0"
numpy = "^1.26.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
